Exercism - Wednesday - 01-18-23

+++ java/need-for-speed/src/main/java/NeedForSpeed.java
@@ -56,7 +56,7 @@ class RaceTrack {
 
     public boolean tryFinishTrack(NeedForSpeed car) {
 
-        if((distance / car.getSpeed()) * car.getBatteryDrain() <= 100)
+        if((double)((double)distance / car.getSpeed()) * car.getBatteryDrain() <= 100)
             return true;
         return false;
